requirements.txt
Django==6.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.10